pyproject.toml
[tool.poetry]
name = "repl_python3_beta_Discord-Chat-Bot"
version = "0.1.0"
description = ""
authors = ["Sas2k <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.19.1"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
